차량 반납 완료 시 관리자 예약취소 버튼 보이지 않게

+++ WebContent/WEB-INF/views/admin/adminTotalReserList.jsp
@@ -61,9 +61,13 @@ String searchKeyword = request.getParameter("searchKeyword");
             		<td><%= "auto".equals(reservation.getLicenseType()) ? "2종" : "1종" %></td>
             		<td><%= reservation.getReviewStatus() %></td>
             		<td><%= reservation.getReturnStatus() %></td>
+<% if ("N".equals(reservation.getReturnStatus())) { %>
             		<td>
             		<input type="button" value="취소" class="adminCarUpdateBtn" onclick="location.href='<%= request.getContextPath() %>/admin/adminReservationDelete?reserNo=<%= reservation.getReserNo() %>';"/>
             		</td>
+<% } else { %>
+					<td></td>
+<% } %>
             	</tr>
 <%		
 	}
